requirements.txt
Django>=4.0
djangorestframework
djangorestframework-gis
psycopg2-binary
GDAL